Add enchanting in tables (#681)
* Entity reworking for spawning

* Cleanup dye item and fix relative entity movement (it has no use now anymore)

* Move glazed terracotta into one API interface (like wool and dye)
Re-enable "relative" movement

* Add new inventory transaction system. Also fix some container related issues (mostly closing them)

* Remove my test output jsons

* Add creative inventory transactions

* Enchanting WIP

* Added enchantment handling. There are some nasty (i guess client sided) bugs which need fixing before this is ready though.

* Fix off by one client sided bug. Implement events needed

* Bring log level back to info

* Fix unit test (placement caused NPE on furnaces)

+++ gomint-server/src/main/java/io/gomint/server/entity/AttributeInstance.java
@@ -25,7 +25,7 @@ public class AttributeInstance {
     private float value;
     private boolean dirty;
 
-    private Map<AttributeModifierType, Map<AttributeModifier, Float>> modifiers = new EnumMap<>( AttributeModifierType.class );
+    private Map<AttributeModifierType, Map<AttributeModifier, Double>> modifiers = new EnumMap<>( AttributeModifierType.class );
 
     AttributeInstance( String key, float minValue, float maxValue, float value ) {
         this.key = key;
@@ -36,14 +36,14 @@ public class AttributeInstance {
         this.dirty = true;
     }
 
-    public void setModifier( AttributeModifier modifier, AttributeModifierType type, float amount ) {
-        Map<AttributeModifier, Float> mods = this.getModifiers( type );
+    public void setModifier( AttributeModifier modifier, AttributeModifierType type, double amount ) {
+        Map<AttributeModifier, Double> mods = this.getModifiers( type );
         mods.put( modifier, amount );
         this.recalc();
     }
 
-    private Map<AttributeModifier, Float> getModifiers( AttributeModifierType type ) {
-        Map<AttributeModifier, Float> modifier = this.modifiers.get( type );
+    private Map<AttributeModifier, Double> getModifiers( AttributeModifierType type ) {
+        Map<AttributeModifier, Double> modifier = this.modifiers.get( type );
         if ( modifier == null ) {
             modifier = new EnumMap<>( AttributeModifier.class );
             this.modifiers.put( type, modifier );
@@ -55,7 +55,7 @@ public class AttributeInstance {
     private void recalc() {
         this.value = this.defaultValue;
 
-        for ( Map.Entry<AttributeModifierType, Map<AttributeModifier, Float>> entry : this.modifiers.entrySet() ) {
+        for ( Map.Entry<AttributeModifierType, Map<AttributeModifier, Double>> entry : this.modifiers.entrySet() ) {
             this.calcModifiers( entry.getKey(), entry.getValue() );
         }
 
@@ -64,24 +64,24 @@ public class AttributeInstance {
         this.dirty = true;
     }
 
-    private void calcModifiers( AttributeModifierType type, Map<AttributeModifier, Float> value ) {
+    private void calcModifiers( AttributeModifierType type, Map<AttributeModifier, Double> value ) {
         switch ( type ) {
             case ADDITION:
-                for ( Float aFloat : value.values() ) {
+                for ( Double aFloat : value.values() ) {
                     this.value += aFloat;
                 }
 
                 break;
 
             case ADDITION_MULTIPLY:
-                for ( Float aFloat : value.values() ) {
+                for ( Double aFloat : value.values() ) {
                     this.value += this.defaultValue * aFloat;
                 }
 
                 break;
 
             case MULTIPLY:
-                for ( Float aFloat : value.values() ) {
+                for ( Double aFloat : value.values() ) {
                     this.value *= 1f + aFloat;
                 }
 
@@ -90,7 +90,7 @@ public class AttributeInstance {
     }
 
     public void removeModifier( AttributeModifier modifier ) {
-        for ( Map.Entry<AttributeModifierType, Map<AttributeModifier, Float>> entry : this.modifiers.entrySet() ) {
+        for ( Map.Entry<AttributeModifierType, Map<AttributeModifier, Double>> entry : this.modifiers.entrySet() ) {
             entry.getValue().remove( modifier );
         }
 
@@ -146,7 +146,7 @@ public class AttributeInstance {
                 }
 
                 int operation = nbtAmplifier.getInteger( "Operation", 0 );
-                float amount = nbtAmplifier.getFloat( "Amount", 0f );
+                double amount = nbtAmplifier.getDouble( "Amount", 0.0 );
 
                 if ( modifier != null && amount != 0 ) {
                     switch ( operation ) {
@@ -177,8 +177,8 @@ public class AttributeInstance {
         // Check for 0 mode multipliers (simple addition)
         List<NBTTagCompound> nbtModifiers = new ArrayList<>();
         if ( !this.modifiers.isEmpty() ) {
-            for ( Map.Entry<AttributeModifierType, Map<AttributeModifier, Float>> entry: this.modifiers.entrySet() ) {
-                for (Map.Entry<AttributeModifier, Float> modifierEntry : entry.getValue().entrySet()) {
+            for ( Map.Entry<AttributeModifierType, Map<AttributeModifier, Double>> entry: this.modifiers.entrySet() ) {
+                for (Map.Entry<AttributeModifier, Double> modifierEntry : entry.getValue().entrySet()) {
                     NBTTagCompound nbtTagCompound = new NBTTagCompound( "" );
                     nbtTagCompound.addValue( "Name", modifierEntry.getKey().getName() );
                     nbtTagCompound.addValue( "Operation", entry.getKey().ordinal() );
@@ -212,7 +212,7 @@ public class AttributeInstance {
         return value;
     }
 
-    public Map<AttributeModifierType, Map<AttributeModifier, Float>> getModifiers() {
+    public Map<AttributeModifierType, Map<AttributeModifier, Double>> getModifiers() {
         return modifiers;
     }
